GCP pub-sub/requirements.txt
google-cloud-pubsub
dist/chronicle-0.0.1-py3-none-any.whl
requests==2.27.1
jwt==1.3.1
google-auth==2.6.0
google-cloud-secret-manager==2.10.0